refactor: http server, snapshot

+++ app/src/main/kotlin/li/songe/gkd/data/BaseSnapshot.kt
@@ -6,7 +6,7 @@ interface BaseSnapshot {
     val appId: String?
     val activityId: String?
     val appName: String?
-    val appVersionCode: Int?
+    val appVersionCode: Long?
     val appVersionName: String?
 
     val screenHeight: Int
